Change to create flow & cleanup, reorganise utility methods etc
